Rename postcss plugin to internal-postcss-fix-relative-paths
