+++ packages/@tailwindcss-cli/tsup.config.ts
@@ -5,5 +5,5 @@ export default defineConfig({
   clean: true,
   minify: true,
   entry: ['src/index.ts'],
-  noExternal: ['postcss-fix-relative-paths-plugin'],
+  noExternal: ['internal-postcss-fix-relative-paths'],
 })
