+++ packages/@tailwindcss-postcss/tsup.config.ts
@@ -8,5 +8,5 @@ export default defineConfig({
   cjsInterop: true,
   dts: true,
   entry: ['src/index.ts'],
-  noExternal: ['postcss-fix-relative-paths-plugin'],
+  noExternal: ['internal-postcss-fix-relative-paths'],
 })
