+++ packages/internal-postcss-fix-relative-paths/src/fixtures/external-import/src/index.css
